Updated directory structure and introduced sample notebook
Changes include an update to directory structure which introduces
the following directories: datasets (to store dataset-indicating .txt files),
features (to store the output of preprocessing), notebooks (to store any and
all jupyter notebooks) and finally outputs (to store pipeline-level outputs).

The included sample notebook -- SampleSparkDeploy.ipynb -- demonstrates a simple
process and procedure for tallying a word count across a specified dataset. This
can serve as infrastructure for further development of both models and the mature
preprocessing script